+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.0] - 2026-06-12
+
+### Added
+
+- Rule for TMDB genres
+  - Added a TMDB genre rule picker modal
+- Rule for media server collections _(does require a full scan)_
+  - Added a media server collection rule picker modal
+- Additional metadata for candidates on all views
+  - Library name
+  - Watch count
+  - Added date
+  - Last viewed
+
+### Changed
+
+- Change 'Date Added` to 'Flagged' for candidates for clarity
+- Updated Granian from 2.7.6 to 2.7.7
+
 ## [0.1.0-beta.30] - 2026-06-09
 
 ### Changed
@@ -16,7 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update HTTP status code from UNPROCESSABLE_ENTITY to UNPROCESSABLE_CONTENT
 - Plex could fail when querying episodes with very large libraries/slow IO, this has been resolved with pagination for: Collections, Episodes, Movies, and Series where applicable
 
-## [0.1.0-beta.30] - 2026-06-04
+## [0.1.0-beta.29] - 2026-06-04
 
 ### Added